buoi6_HuynhDangKhoa/requirements.txt
mysql-connector-python==8.0.33
lxml==4.9.3